Fixes error where app module cannot be imported
When calling `module()` with `inject()`, an error is thrown:
```
Error: [$injector:modulerr] Failed to instantiate module ng due to:
TypeError: 'undefined' is not an object (evaluating
'Function.prototype.bind.apply')
```

In karma-phantomjs-launcher 1.x, `Function.prototype.bind` isn't defined. Upgrading to 2.x
fixes this problem. In addition, it requires phantomjs-prebuilt as a
peer dependency.

(As an aside, the test was masking over the error. A more specific
error, or a message assertion would be better practice).